requirements.txt
torch
torchvision
torchaudio
numpy
scipy
matplotlib
pandas
pillow
pyyaml
requests
tqdm
click
easydict
lpips
progressbar